feat : MQTT subscribe and update MU and CU

+++ src/main/kotlin/measuremanager/settingsmanager/dtos/CommandDTO.kt
@@ -1,3 +1,3 @@
 package measuremanager.settingsmanager.dtos
 
-data class CommandDTO(val commandId : Long, val gateway : Int, val cu : Int, val mu: Int, val type: String, val cuSettingDTO: CuSettingDTO?, val muSettingDTO: MuSettingDTO?  )
+data class CommandDTO(val commandId : Long, val gateway : Long, val cu : Long, val mu: Long, val type: String, val cuSettingDTO: CuSettingDTO?, val muSettingDTO: MuSettingDTO?  )
